% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ADPvars}
\alias{ADPvars}
\title{External variables for the respondents in the SUPwide and ADPwide response
data set}
\format{A data frame with 38 rows and 8 columns
\describe{
  \item{q1}{A categorical variable indicating the subject the respondent
            taught in the 13-14 school year}
  \item{ELA.SS}{A string variable indicating if the respondent is an English-
                Language Arts or Social Studies teacher}
  \item{CORE}{A string variable indicating if the respondent teaches a 'core'
              (state-tested) subject}
  \item{q2}{The number of years the respondent has taught}
  \item{experience}{A string indicating if the teacher has 1-5 years of
                    experience or more ("6+")}
  \item{q3}{The number of years the respondent has taught in the district}
  \item{tenure.Y.N}{A string indicating if the teacher has tenure}
  \item{only.job}{A string indicating if the teacher's current job was the
                  only teaching job he or she has held}
}}
\source{
Amy Arneson's Measurement 1 study, collected in Fall 2013
}
\usage{
ADPvars
}
\description{
Values of external variables that were measured in the SUP/ADP study
}
\keyword{datasets}
