% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ADPwide}
\alias{ADPwide}
\title{Response data for the Adoption study}
\format{A data frame with 38 rows and 13 columns.
\describe{
  The row names are the respondent IDs.  The column names are the items.  The
  values are the score received.  All items were scored polytomously.
}}
\source{
Amy Arneson's Measurement 1 study, collected in Fall 2013
}
\usage{
ADPwide
}
\description{
The response data (scores) for the items aligned to the Adoption construct
}
\keyword{datasets}
