% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{AMYcons}
\alias{AMYcons}
\title{Construct information for the combined ADP/SUP study}
\format{A data frame with 2 rows and 8 columns:
\describe{
  \item{cons.ID}{A unique numerical ID to map items to constructs}
  \item{long.name}{The name of the construct}
  \item{short.name}{An abbreviation of the construct name}
  \item{cat1}{The name of the first category}
  \item{cat2}{The name of the second category}
  \item{cat3}{The name of the third category}
  \item{cat4}{The name of the fourth category}
  \item{cat5}{The name of the fifth category}
}}
\source{
Amy Arneson's Measurement 1 study, collected in Fall 2013
}
\usage{
AMYcons
}
\description{
The required construct information for the Adoption & Support data
}
\keyword{datasets}
