% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{AMYwide}
\alias{AMYwide}
\title{Response data for the combined ADP/SUP study}
\format{A data frame with 38 rows and 28 columns.
\describe{
  The row names are the respondent IDs.  The column names are the items.  The
  values are the score received.  All items were scored polytomously.
}}
\source{
Amy Arneson's Measurement 1 study, collected in Fall 2013
}
\usage{
AMYwide
}
\description{
The response data (scores) for the items aligned to the Support construct
}
\keyword{datasets}
