% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{CPC.graph}
\alias{CPC.graph}
\title{Plots the cumulative probability curves from \code{craschR} output.}
\usage{
CPC.graph(results, itemOrder = NULL, palette = "BASS", observed = FALSE,
  minCell = 8, focusTheta = NULL, logitAxis = "x", writeout = FALSE,
  imageType = "pdf", fileSuffix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{itemOrder}{A numeric vector that specifies which items from the output
should be graphed.  If \code{NULL}, all items will be graphed.}

\item{palette}{A character string indicating the color scheme.  Can be
"BASS", "grey", or any RColorBrewer palette.}

\item{observed}{A logical indicating whether or not the observed proportions
should be graphed as points along with the curve.}

\item{minCell}{A positive integer indicating the smallest cell value to plot
an observed proportion.  Only applies if \code{observed=TRUE}.}

\item{focusTheta}{A numeric vector indicating at which logit values to plot
a vertical line and print the probabilities for each category.  Only
applies if \code{observed=FALSE}.}

\item{logitAxis}{A character indicating if the logits should be graphed on
the x-axis (traditional) or y-axis (as in Wilson, 2005).  Choose from
c("x", "y").}

\item{writeout}{A logical indicating whether the graphic should be written to
to your working directory as your specified \code{imageType}.  If
\code{TRUE}, the file name will begin \code{CPC}, an index, and the
\code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
One plot for each item is created.
}
\description{
Plots the cumulative probability curves from \code{craschR} output.
}

