% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crasch.R
\docType{package}
\name{crasch}
\alias{crasch}
\alias{crasch-package}
\title{crasch: A package developed to accompany an introductory measurement course
  using Wilson (2005).}
\description{
An organizational and analysis tool that faciliatates implemenation of the
  Berkeley Assessment System, following that of Wilson (2004). Provides
  user-friendly wrappers for Rasch model estimation (performed by either mirt
  or TAM). Produces nicely formatted and readable tables and graphics for
  with little to no R-coding experience.
}
\section{Author}{

  Amy E. Arneson (aea@berkeley.edu), Maintainer
}

\section{Contributors}{

  Mark Wilson,
  Rebecca Freund,
  David Torres Irribarra,
  Yukie Toyama
}

\section{Citation note}{

  Be sure to cite this package, the estimation package used (TAM or mirt),
  and the WrightMap package (if you use the Wright Map graphic).
}

