% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craschR.R
\name{craschR}
\alias{craschR}
\title{Item and person estimates and fit statistics.}
\usage{
craschR(scores, itemInfo = NULL, consInfo = NULL, varsInfo = NULL,
  estPackage = "TAM", retainOrig = FALSE, missingAs0 = FALSE,
  longFormat = FALSE, persMethod = "EAP", consecutive = FALSE,
  writeout = TRUE, fileSuffix = NULL)
}
\arguments{
\item{scores}{A data frame or path to CSV file in crasch format with the
scoring information. The scores are be coded as integers in terms of the
construct levels. Missing responses are coded as \code{NA}.}

\item{itemInfo}{A data frame or path to CSV file in crasch format with
information on each item. There must be \code{I} rows and as many columns
as specified in the crasch format documentation. No factor variables are
allowed.}

\item{consInfo}{A data frame or path to CSV file containing labeling
information about the construct(s) in crasch format. Contains 1 row per
construct.}

\item{varsInfo}{A data frame containing demographic variable information for
respondents. One respondent per row, IDs must match those of
\code{scores} data frame. Numeric, character, and factor variables are
allowed.}

\item{estPackage}{A string that identifies which IRT package will be used to
conduct estimation of item (and possibly person) parameters. See details.}

\item{retainOrig}{A logical indicating whether one element of the output
object should include the original result object from \code{mirt} or
\code{TAM}.}

\item{missingAs0}{A logical indicating whether \code{NA}s should be recoded
to the lowest level (available on a given item) of the construct.
If \code{FALSE}, missing data will remain as missing data.}

\item{longFormat}{A logical indicating whether the scores object is in long
format.  If \code{TRUE}, reshaping to wide format will occur internally.}

\item{persMethod}{Choose from c("EAP", "MLE", "WLE")}

\item{consecutive}{A logical indicating whether to perform a 'consecutive'
analysis in which an independent analysis is conducted for each dimension.
If \code{TRUE}, the correlation between any two dimensions will be
constrained to be 0. If \code{FALSE}, multidimensional analysis that
allows non-zero correlations among dimensions will be conducted.}

\item{writeout}{A logical indicated whether the estimate objects should be
written to your working directory as CSVs.  See Return below for the names
of each file produced.  Each will also include any \code{fileSuffix} that
you specify.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
A list with the following entries
  \item{itemPars}{A matrix with the item parameter estimates.  If
  \code{estPackage = "mirt"}, this matrix will be...
  \code{estPackage = "TAM"}, this matrix will be in the ConQuest
  parameterization with an overall item `delta' and step `tau's.}
  (\code{itemGrid.csv})
  \item{itemSEs}{A matrix with the standard error estimates corresponding to
  the estimates given in \code{itemPars}.} (\code{itemGrid.csv})
  \item{itemThres}{A matrix giving the Thurstonian threshold estimates.}
  (\code{itemGrid.csv})
  \item{itemFit}{A data frame providing the fit statistics for items and
  steps.} (\code{itemFit.csv})
  \item{persPars}{A data frame containing estimates for person locations for
  each dimension.} (\code{persGrid.csv})
  \item{persSEs}{A data frame with standard errors for person estimates.}
  (\code{persGrid.csv})
  \item{persFit}{If non-consecutive (including unidimensional) analysis is
  performed, this will be a data frame with person fit statistics.  If
  `consecutive' multidimensional analysis is performed (independent analysis
  of each dimension), this will be a list with fit statistics for each
  dimension.  True multidimensional fit statistics are not available if
  \code{TAM} is used.  See details for how these are produced.}
  (\code{persFit.csv})
  \item{popDist}{A list containing a vector of estimated population means
  for each dimension (often these are constrained to be 0) and a variance-
  covariance matrix.  Variances appear on the diagonal, covariances
  elsewhere.  `Consecutive' analyses fix all covariances to 0.}
  (\code{popDist.csv})
  \item{sepRel}{A vector containing the separation reliabilities for each
  dimension.} (\code{classicalStats.csv})
  \item{estSummary}{A list containing information about the estimation
  settings and model fit.  Use the \code{names()} function to see what is
  provided here.} (\code{estSummary.csv})
  \item{classicalStats}{A list containing information about each dimension.
  Notably, Cronbach's Alpha and person separation reliabilities are
  provided.  Statistics are calculated by recoding all missings to 0 as well
  as by only including complete cases.} (\code{classicalStats.csv})
  \item{empties} A list of vectors (1 per item) of the categories (if any)
  that were deemed possible, but no one scored into them.
  \item{estResults}{If \code{retainOrig = TRUE}, this will contain the FULL
  output provided by the estimation package.  See the documentation for that
  package for full details.  Otherwise, this will be NULL.}
  \item{scoresOrig}{}
  \item{scoresRecoded}{Wide format score data, recoded to nonnegative
  successive integers.}
  \item{itemInfo}{}
  \item{consInfo}{}
  \item{varsInfo}{}
}
\description{
Item and person estimates and fit statistics.
}
