% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{ICC.graph}
\alias{ICC.graph}
\title{Plots the item characteristic curves (or category characteristic curves, if
polytomous) from \code{craschR} output.}
\usage{
ICC.graph(results, itemOrder = NULL, palette = "BASS", writeout = FALSE,
  imageType = "pdf", fileSuffix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{itemOrder}{A numeric vector that specifies which items from the output
should be graphed.  If \code{NULL}, all items will be graphed.}

\item{palette}{A character string indicating the color scheme.  Can be
"BASS", any RColorBrewer palette (the 3-color palette will be used), or a
vector containing 3 colors.}

\item{writeout}{A logical indicating whether the graphic should be written to
to your working directory as your specified \code{imageType}.  If
\code{TRUE}, the file name will begin \code{ICC}, an index, and the
\code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
One plot for each item is created.
}
\description{
Plots the item characteristic curves (or category characteristic curves, if
polytomous) from \code{craschR} output.
}
