% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{infit.MNSQ}
\alias{infit.MNSQ}
\title{Plots in infit mean sqare statistics from \code{craschR} output.}
\usage{
infit.MNSQ(results, itemOrder = NULL, params = "items", palette = "BASS",
  writeout = FALSE, imageType = "pdf", fileSuffix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{itemOrder}{A numeric vector that specifies which items from the output
should be graphed.  If \code{NULL}, all items will be graphed.}

\item{params}{A string indicating for which parameters fit should be graphed.
Can be \code{"items"} or \code{"steps"}.}

\item{palette}{A character string indicating the color scheme.  Can be
"BASS", any RColorBrewer palette (the first 2 of the 3-color palette will
be used), or a vector containing 2 colors: the first for the shading of the
"acceptable" region and the second for the point color.}

\item{writeout}{A logical indicating whether the graphic should be written to
to your working directory as your specified \code{imageType}.  If
\code{TRUE}, the file name will begin \code{infit}, followed by the
\code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
One or more plots of the infit mean squares with "acceptable" region
  highlighted.
}
\description{
Plots in infit mean sqare statistics from \code{craschR} output.
}
