% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{item.scores}
\alias{item.scores}
\title{Create a stacked bar graph showing score level frequencies per item.}
\usage{
item.scores(results, dim = NULL, freqs = TRUE, palette = "Set3",
  writeout = FALSE, imageType = "pdf", filePrefix = NULL, ...)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{dim}{Specify which dimension(s) to create graphic/tables for.  If
\code{NULL}, output and graphics for each dimension will be produced.}

\item{freqs}{A logical indicated whether frequencies or raw counts should be
graphed.}

\item{palette}{Can be "blue", "grey", or any RColorBrewer palette string.}

\item{writeout}{A logical indicated whether the estimate objects should be
written to your working directory as CSVs.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{filePrefix}{A character string that will be affixed to the beginning
of each file (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}

\item{...}{Additional arguments to be passed to \code{barplot} function.}
}
\value{
Creates a stacked bar plot and returns a list with the following
  entries:
  \item{counts}{A matrix with the raw counts of scores given to each
  construct level on each item.  \code{NA}s indicate that the score level
  was not possible while 0s indicate that the score level was on the outcome
  space yet no one scored at that level.}
  \item{proportions}{A matrix with proportions of scores given to each
  construct level on each item.}
  Note that there will be one matrix per dimension.  If a multi-dimensional
  analysis was run
}
\description{
Create a stacked bar graph showing score level frequencies per item.
}

