% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{KIDMAP}
\alias{KIDMAP}
\title{Plots a KIDMAP for a given person and construct.}
\usage{
KIDMAP(results, personID, dim = NULL, probBounds = c(0.25, 0.75),
  palette = "BASS", writeout = FALSE, imageType = "pdf",
  fileSuffix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{personID}{A character or numeric vector indicating the index ID of the
respondent for the plot.  (Do not specify a row number, use the assigned
ID.)}

\item{dim}{A numeric vector that specifies for which dimension(s) to create
graphic/tables.  If \code{NULL}, output and graphics for each dimension
will be produced.}

\item{probBounds}{A vector containing the range of probabilities which are
"unsurprising."}

\item{palette}{The color scheme for the plot. Can be "BASS or any
RColorBrewer palette name (the 3-color palette will be used). Can also
specify a vector with 3 colors in any R-supported form; the first is the
color of the "expected" area, the second the "surprise" area, and the third
the point color.}

\item{writeout}{A logical indicating whether the graph should be written to
your working directory as your indicated \code{imageType}.  If \code{TRUE},
the file name will begin \code{KIDMAP} and will include an index for the
selected respondent, and the \code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
KIDMAP plot for specified person.
}
\description{
Plots a KIDMAP for a given person and construct.
}
