% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{mn.traj}
\alias{mn.traj}
\title{Returns and plots the mean location trajectories by category per item.}
\usage{
mn.traj(results, itemOrder = NULL, palette = "BASS", writeout = FALSE,
  imageType = "pdf", fileSuffix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{itemOrder}{A numeric vector that specifies which items from the output
should be graphed.  If \code{NULL}, all items will be graphed.  Note that
items scored on different dimensions will be graphed separately.}

\item{palette}{A character string indicating the color scheme.  Can be
"BASS", any RColorBrewer palette or a vector containing any number of
 colors.  The colors from each palette will be circulated through items.}

\item{writeout}{A logical indicating whether the graphic should be written to
to your working directory as your specified \code{imageType}.  If
\code{TRUE}, the file name will begin \code{meanTraj}, followed by an index
for the dimension, and the \code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
A list containing matrices of the mean location estimates by score
  and their standard deviations.
}
\description{
Returns and plots the mean location trajectories by category per item.
}

