% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{Sp.rho}
\alias{Sp.rho}
\title{Calculates Spearman's rho for fully dichotomous, unidimensional data.}
\usage{
Sp.rho(results, whichStep = 1, expOrder)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{whichStep}{An integer indicating which step you want to compare (only
applicable if you have 'equivalently scored' polytomous items meaning that
all items are scored into the same levels and all scores are present in
your data).}

\item{expOrder}{A vector with the expected ordering of the items.}
}
\value{
The split-halves reliability coefficient.
}
\description{
Calculates Spearman's rho for fully dichotomous, unidimensional data.
}

