% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{Split.halves}
\alias{Split.halves}
\title{Calculates the split halves reliability coefficient.}
\usage{
Split.halves(results1, results2)
}
\arguments{
\item{results1}{The output from a run of \code{craschR}. (link?)}

\item{results2}{The output from a run of \code{craschR}.  The estimation
must be run on the same respondents with the same estimation settings.  The
only difference in the two runs of \code{craschR} should be the items used.
Note that you can only include items scored on the same dimension.}
}
\value{
The split-halves reliability coefficient.
}
\description{
Calculates the split halves reliability coefficient.
}
