% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{SUPitem}
\alias{SUPitem}
\title{Item information for the SUPwide response data set}
\format{A data frame with 15 rows and 10 columns
\describe{
  \item{item.ID}{A unique numerical ID for the item}
  \item{item.name}{The name of the item}
  \item{cons.ID}{The unique numerical construct ID that maps the construct to
                 the item}
  \item{item.type}{A character string indicating the type of the item (This
                   column is only used internally in BASS)}
  \item{fixed}{A logical indicating if the item has a fixed parameter(s)
               (This is only to be used internally in BASS)}
  \item{cat1}{A logical indicating if the category is present in the scoring
              guide for the item}
  \item{cat2}{A logical indicating if the category is present in the scoring
              guide for the item}
  \item{cat3}{A logical indicating if the category is present in the scoring
              guide for the item}
  \item{cat4}{A logical indicating if the category is present in the scoring
              guide for the item}
  \item{cat5}{A logical indicating if the category is present in the scoring
              guide for the item}
}}
\source{
Amy Arneson's Measurement 1 study, collected in Fall 2013
}
\usage{
SUPitem
}
\description{
The required item information for the Support data
}
\keyword{datasets}

