% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{item.scores}
\alias{item.scores}
\title{Create a stacked bar graph showing score level frequencies per item.}
\usage{
item.scores(results, dim = NULL, freqs = TRUE, palette = "BASS",
  writeout = FALSE, imageType = "pdf", fileSuffix = NULL, ...)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{dim}{A numeric vector that specifies for which dimension(s) to create
graphic/tables.  If \code{NULL}, output and graphics for each dimension
will be produced.}

\item{freqs}{A logical indicated whether frequencies or raw counts should be
graphed.}

\item{palette}{A character string indicating the color scheme for the graph.
Can be "BASS", "grey", or any RColorBrewer palette string.}

\item{writeout}{A logical indicating whether the graphics/tables should be
written to to your working directory as your specified \code{imageType} and
CSVs.  If \code{TRUE}, the file name will begin \code{itemscores} and will
include an index (if more than one graph is produced) and the
\code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}

\item{...}{Additional arguments to be passed to \code{barplot} function.}
}
\value{
Creates a stacked bar plot and returns a list with the following
  entries:
  \item{counts}{A matrix with the raw counts of scores given to each
  construct level on each item.  \code{NA}s indicate that the score level
  was not possible while 0s indicate that the score level was on the outcome
  space yet no one scored at that level.}
  \item{proportions}{A matrix with proportions of scores given to each
  construct level on each item.}
  Note that there will be one matrix per dimension.  If a multi-dimensional
  analysis was run
}
\description{
Create a stacked bar graph showing score level frequencies per item.
}
