% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVisualizations.R
\name{wm}
\alias{wm}
\title{A wrapper for the wrightMap() function from the WrightMap package.}
\usage{
wm(results, dim = NULL, itemOrder = "item", palette = "BASS",
  writeout = FALSE, imageType = "pdf", filePrefix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{dim}{A numeric vector to specify which dimension(s) to create
graphic/tables for.  If \code{NULL}, output and graphics for each dimension
will be produced.}

\item{itemOrder}{Can be "item", "construct", or a numeric vector.  If "item",
items will be ordered as in \code{itemInfo}.  If "construct", items will
be grouped by the levels of the construct.}

\item{palette}{Color scheme.  Can be "BASS" (default) or any RColorBrewer
palette.  If you want to customize further, the wrightMap() function should
be used directly, not this wrapper.}

\item{writeout}{A logical indicated whether the estimate objects should be
written to your working directory as CSVs.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{filePrefix}{A character string that will be affixed to the beginning
of each file (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
Wright Map graphic
}
\description{
A wrapper for the wrightMap() function from the WrightMap package.
}

