% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personItem.R
\name{wm}
\alias{wm}
\title{A wrapper for the wrightMap() function from the WrightMap package.}
\usage{
wm(results, dim = NULL, byCat = FALSE, palette = "BASS",
  writeout = FALSE, imageType = "pdf", fileSuffix = NULL)
}
\arguments{
\item{results}{The output from a run of \code{craschR}. (link?)}

\item{dim}{A numeric vector that specifies for which dimension(s) to create
graphic/tables.  If \code{NULL}, output and graphics for each dimension
will be produced.}

\item{byCat}{A logical indicating if items/steps should be grouped by the
construct category.  If \code{FALSE}, items will be plotted in the order
of \code{itemInfo}.}

\item{palette}{A character string indicating the color scheme.  Can be "BASS"
or any RColorBrewer palette.  If you want to customize further, the
\code{wrightMap()} function should be used directly, not this wrapper.}

\item{writeout}{A logical indicating whether the graphic should be written to
to your working directory as your specified \code{imageType}.  If
\code{TRUE}, the file name will begin \code{WM} and will include an index
(if more than one graph is produced) and the \code{fileSuffix} if provided.}

\item{imageType}{A character string indicating the format for graphics (if
\code{writeout = TRUE}). Supported types:
\code{c("pdf","svg","jpeg","bmp","tiff","png")}.}

\item{fileSuffix}{A character string that will be affixed to the end of each
file name (if \code{writeout = TRUE}). Use this if you are conducting
multiple analyses in the same working directory and do not wish for your
existing files to be overwritten.}
}
\value{
Wright Map graphic (multiple graphics will be produced if you have
  more than one dimension and ran consecutive analyses)
}
\description{
A wrapper for the wrightMap() function from the WrightMap package.
}
